requirements.txt
pyjq
requests
python-magic
python-dotenv